Parallelise image updating
Images can be operated on in parallel, allowing the updating of the
individual images (especially with the latest cumulative update) to be
done in parallel. The dism tool appears to be single-threaded for these
operations, so the speed-up is pretty good.

A tiny bit more speed can be squeezed out by updating boot.wim at the
same time as install.wim - this slightly complicates the script because
the mount index and mount-directory index won't match (this could be
fixed by having multiple prefixes, or just by coping). Technically,
lang.ini needs to have been updated before the install.wim locales are
changed (this is academic, as it's inconceivable that the boot.wim
updates would take longer than install.wim).

Signed-off-by: David Allsopp <david.allsopp@metastack.com>